🔧 Refactor: Update import paths to use new _dependencies structure and add user_groups functionality
