+++ services/catalog/src/simcore_service_catalog/api/_dependencies/director.py
