+++ services/catalog/src/simcore_service_catalog/core/background_tasks.py
@@ -20,7 +20,7 @@ from models_library.services import ServiceMetaDataPublished
 from models_library.services_types import ServiceKey, ServiceVersion
 from packaging.version import Version
 from pydantic import ValidationError
-from simcore_service_catalog.api.dependencies.director import get_director_api
+from simcore_service_catalog.api._dependencies.director import get_director_api
 from simcore_service_catalog.service import manifest
 from sqlalchemy.exc import SQLAlchemyError
 from sqlalchemy.ext.asyncio import AsyncEngine
