+++ services/catalog/tests/unit/test_service_manifest.py
@@ -13,7 +13,7 @@ from models_library.function_services_catalog.api import is_function_service
 from pytest_simcore.helpers.monkeypatch_envs import setenvs_from_dict
 from pytest_simcore.helpers.typing_env import EnvVarsDict
 from respx.router import MockRouter
-from simcore_service_catalog.api.dependencies.director import get_director_api
+from simcore_service_catalog.api._dependencies.director import get_director_api
 from simcore_service_catalog.infrastructure.director import DirectorApi
 from simcore_service_catalog.service import manifest
